Moved getting phrases from card to app
